requirements.txt
numpy
pandas
Pillow
tqdm
joblib
scikit-learn
torch
torchvision
kagglehub
kaggle
lightgbm
matplotlib
